# specs/001-build-a-web/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Chain Crypto Wallet and Exchange Platform API
  description: |
    RESTful API for cryptocurrency wallet management, transactions, and exchange operations.
    Supports Bitcoin (BTC), Ethereum (ETH), Solana (SOL), and Stellar (XLM) blockchains.
  version: 1.0.0
  contact:
    name: API Support
    email: api@cryptowallet.example.com
  license:
    name: Proprietary
    url: https://cryptowallet.example.com/license
servers:
  - url: https://api.cryptowallet.example.com/v1
    description: Production server
  - url: https://staging-api.cryptowallet.example.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /wallets/{id}:
    get:
      tags: [Wallets]
      summary: Get wallet details
      operationId: getWallet
      parameters:
        - $ref: '#/components/parameters/WalletId'
      responses:
        '200':
          description: Wallet retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /wallets/{id}/transactions:
    get:
      tags: [Wallets]
      summary: Get wallet transactions
      operationId: getWalletTransactions
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  # ==================== Transactions ====================
  /transactions:
    get:
      tags: [Transactions]
      summary: List user transactions
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: chain
          in: query
          schema:
            $ref: '#/components/schemas/BlockchainChain'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Transactions]
      summary: Create and broadcast transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /exchange/quote:
    post:
      tags: [Exchange]
      summary: Get exchange quote
      operationId: getQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Quote generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequest'
  /exchange/swap:
    post:
      tags: [Exchange]
      summary: Execute token swap
      operationId: executeSwap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapRequest'
      responses:
        '201':
          description: Swap executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /kyc/submit:
    post:
      tags: [KYC]
      summary: Submit KYC profile
      operationId: submitKYC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KYCSubmitRequest'
      responses:
        '201':
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
  /kyc/documents:
    post:
      tags: [KYC]
      summary: Upload verification document
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [document_type, file]
              properties:
                document_type:
                  type: string
                  enum: [passport, national_id, drivers_license, proof_of_address, selfie]
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCDocument'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /analytics/performance:
    get:
      tags: [Analytics]
      summary: Get portfolio performance
      operationId: getPerformance
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: ['24h', '7d', '30d', '90d', '1y', 'all']
      responses:
        '200':
          description: Performance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioPerformance'

  # ==================== Health ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
  parameters:
    WalletId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    BlockchainChain:
      type: string
      enum: [BTC, ETH, SOL, XLM]
    TransactionStatus:
      type: string
      enum: [pending, confirming, confirmed, failed, cancelled]
    TransactionType:
      type: string
      enum: [send, receive, swap_in, swap_out]
    VerificationLevel:
      type: string
      enum: [unverified, basic, full]
    KYCStatus:
      type: string
      enum: [not_started, pending, under_review, approved, rejected, expired]

    # ========== Authentication ==========
...
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        address:
          type: string
        label:
          type: string
        balance:
          type: string
          example: "1.5"
        balance_usd:
          type: string
          example: "50000.00"
        status:
          type: string
          enum: [active, archived]
        created_at:
          type: string
          format: date-time
    WalletDetail:
      allOf:
        - $ref: '#/components/schemas/Wallet'
        - type: object
          properties:
            derivation_path:
              type: string
            balance_updated_at:
              type: string
              format: date-time
            transaction_count:
              type: integer
...
    CreateTransactionRequest:
      type: object
      required: [wallet_id, to_address, amount]
      properties:
        wallet_id:
          type: string
          format: uuid
        to_address:
          type: string
        amount:
          type: string
        fee_level:
          type: string
          enum: [slow, standard, fast]
          default: standard
        memo:
          type: string
          maxLength: 500
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        tx_hash:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: string
        fee:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        from_address:
          type: string
        to_address:
          type: string
        confirmations:
          type: integer
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
...
    QuoteRequest:
      type: object
      required: [from_wallet_id, to_chain, amount]
      properties:
        from_wallet_id:
          type: string
          format: uuid
        to_chain:
          $ref: '#/components/schemas/BlockchainChain'
        amount:
          type: string
    Quote:
      type: object
      properties:
        quote_id:
          type: string
          format: uuid
        from_symbol:
          type: string
        to_symbol:
          type: string
        from_amount:
          type: string
        to_amount:
          type: string
        exchange_rate:
          type: string
        fee_percentage:
          type: string
        fee_amount:
          type: string
        expires_at:
          type: string
          format: date-time
    SwapRequest:
      type: object
      required: [quote_id, from_wallet_id, to_wallet_id]
      properties:
        quote_id:
          type: string
          format: uuid
        from_wallet_id:
          type: string
          format: uuid
        to_wallet_id:
          type: string
          format: uuid
    ExchangeOperation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from_symbol:
          type: string
        to_symbol:
          type: string
        from_amount:
          type: string
        to_amount:
          type: string
        exchange_rate:
          type: string
        fee_amount:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        created_at:
          type: string
          format: date-time
        executed_at:
          type: string
          format: date-time

    # ========== KYC ==========
    KYCSubmitRequest:
      type: object
      required: [first_name, last_name, date_of_birth, nationality]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        nationality:
          type: string
          minLength: 2
          maxLength: 2
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            postal_code:
              type: string
            country:
              type: string
    KYCProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        verification_level:
          $ref: '#/components/schemas/VerificationLevel'
        status:
          $ref: '#/components/schemas/KYCStatus'
        submitted_at:
          type: string
          format: date-time
        approved_at:
          type: string
          format: date-time
        daily_limit_usd:
          type: string
        monthly_limit_usd:
          type: string
    KYCDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document_type:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        uploaded_at:
          type: string
          format: date-time

    # ========== Rates ==========
...
    PortfolioPerformance:
      type: object
      properties:
        period:
          type: string
        initial_value_usd:
          type: string
        final_value_usd:
          type: string
        gain_loss_usd:
          type: string
        gain_loss_percentage:
          type: string
        data_points:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              value_usd:
                type: string

    # ========== Common ==========
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...